requirements.txt
PyYAML
tweepy
python-dotenv
bs4
requests
